core: Refactor the code, move FormatResult, LintResult into crate::result mod.
