+++ autocorrect-derive/Cargo.toml
@@ -6,7 +6,7 @@ keywords = ["autocorrect", "lint", "format"]
 license = "MIT"
 name = "autocorrect-derive"
 readme = "../README.md"
-version = "0.1.0"
+version = "0.2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
