+++ autocorrect/src/lib.rs
@@ -94,6 +94,7 @@ mod diff;
 mod format;
 mod fullwidth;
 mod halfwidth;
+mod result;
 mod serde_any;
 mod strategery;
